Moved private repo info out of popover
